updated to bigdecimal

+++ src/main/java/com/ihomziak/clientaccountms/dao/AccountRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.query.Procedure;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -27,6 +28,6 @@ public interface AccountRepository extends JpaRepository<Account, Long> {
     String transferMoney(
             @Param("senderId") String senderUuid,
             @Param("recipientId") String receiverUuid,
-            @Param("amount") double amount
+            @Param("amount") BigDecimal amount
     );
 }
